+++ packages/purplet/src/internal.ts
@@ -3,7 +3,7 @@
 
 import './index';
 
-export * from './lib/gateway';
 export * from './lib/env';
+export * from './lib/gateway';
 export * from './utils/feature';
 export * from './utils/is-directly-run';