+++ packages/purplet/src/lib/env.ts
@@ -1,8 +1,7 @@
 // @ts-expect-error idk, even with installing types this broke.
 import structuredClone from '@ungap/structured-clone';
-import { watch } from 'chokidar';
-import { config } from 'dotenv';
 import path from 'path';
+import { config } from 'dotenv';
 
 const envFile = path.resolve(process.cwd(), '.env');